+++ src/categories/categories-import-from-json/utils/extract-categories.ts
