+++ src/json-file-service/json-write/json.write.module.ts
@@ -5,11 +5,7 @@ import { BrandsModule } from 'src/brands/brands.module';
 import { ReadFileModule } from '../json-read/json.read.module';
 
 @Module({
-  imports: [
-    forwardRef(() => CategoriesModule),
-    forwardRef(() => BrandsModule),
-    ReadFileModule,
-  ],
+  imports: [CategoriesModule, BrandsModule, ReadFileModule],
   providers: [WriteFileService],
   exports: [WriteFileService],
 })
